@@ -25,7 +25,6 @@
     "autoprefixer": "10.4.16",
     "axios": "^0.21",
     "bootstrap": "^4.6.2",
-    "bootstrap-vue": "^2.23.1",
     "cross-env": "^5.1",
     "d3": "^3.5.17",
     "del": "^6.1.1",
@@ -43,20 +42,19 @@
     "replace-in-file-webpack-plugin": "^1.0.6",
     "sass": "^1.68.0",
     "sass-loader": "^11.0.1",
-    "v-tooltip": "^2.0.3",
-    "vue": "^2.6.11",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.10",
+    "tiny-emitter": "^2.1.0",
+    "vue": "^3.3.8",
+    "vue-loader": "^17.3.1",
     "vuedraggable": "^2.24.3",
     "webpack": "^5.89.0"
   },
   "devDependencies": {
-    "@vue/test-utils": "^1.3.6",
-    "@vue/vue2-jest": "^27.0.0",
+    "@vue/test-utils": "^2.4.1",
+    "@vue/vue3-jest": "^29.2.6",
     "axios-mock-adapter": "^1.22.0",
     "babel-core": "^7.0.0-bridge.0",
     "expect": "^27.3.1",
-    "jest": "^27.3.1",
+    "jest": "^29.7.0",
     "moment": "^2.29.4",
     "postcss": "^8.4.31",
     "protractor": "^7.0.0",
