@@ -7,6 +7,7 @@
 </template>
 
 <script>
+import ApiLoader from '../shared/ApiLoader';
 export default {
   name: "HeaderLogo",
 
@@ -18,7 +19,7 @@ export default {
   },
 
   mounted() {
-    this.$root.$on('api-loaded', cdash => {
+    ApiLoader.$on('api-loaded', cdash => {
       if (cdash.logoid > 0) {
         this.logoUrl = `${this.$baseURL}/image/${cdash.logoid}`;
       }
